% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tides.R
\name{get_tides}
\alias{get_tides}
\title{Returns high and low tide information for a specified Station ID}
\usage{
get_tides(api_key, station_id, duration = 1)
}
\arguments{
\item{api_key}{a valid Admirality API Key. The function will load in your API key.}

\item{station_id}{a valid station ID. Can be found by using \code{get_tidal_stations()} or \code{find_closest_tide_station()}}

\item{duration}{duration in days over which to return tide data. Must be >=1 or <=7.}
}
\description{
Returns a \code{tibble} containing data on high and low tides for the specified duration (1 - 7 days)
}
