% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{save_api_key}
\alias{save_api_key}
\title{Register an API key to make requests}
\usage{
save_api_key(api_key)
}
\arguments{
\item{api_key}{a valid Admirality API Key}
}
\description{
Saves an RDS file in the current working directory to store API credentials.
All API-dependent functions will fail until you have done this.
Register for a personal API key here: https://admiraltyapi.portal.azure-api.net/
}
